Move VideoDetails from public api to cli-only
